v. 0.9.7 global refactor and fix work connection with DB
