--- lib/core/domain/models/photosalon.dart
@@ -1,8 +1,0 @@
-import 'package:technical_support_artphoto/core/domain/models/technic.dart';
-
-class Photosalon{
-  String name;
-  List<Technic> technics = [];
-
-  Photosalon(this.name);
-}
