--- lib/core/domain/models/repair.dart
@@ -1,8 +1,0 @@
-import 'package:technical_support_artphoto/core/domain/models/technic.dart';
-
-class Repair{
-  String name;
-  List<Technic> technics = [];
-
-  Repair(this.name);
-}
