--- lib/core/domain/models/storage.dart
@@ -1,8 +1,0 @@
-import 'package:technical_support_artphoto/core/domain/models/technic.dart';
-
-class Storage{
-  String name;
-  List<Technic> technics = [];
-
-  Storage(this.name);
-}
