--- lib/core/domain/models/technic.dart
@@ -1,10 +1,0 @@
-class Technic{
-  int id;
-  int number;
-  String category;
-  String name;
-  String status;
-  String dislocation;
-
-  Technic(this.id, this.number, this.category, this.name, this.status, this.dislocation);
-}
