--- lib/core/domain/models/user.dart
@@ -1,6 +1,0 @@
-class User{
-  final String name;
-  final String access;
-
-  User(this.name, this.access);
-}
