+++ lib/features/auth/presentation/page/authorization.dart
@@ -1,0 +1,189 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:provider/provider.dart';
+import 'package:technical_support_artphoto/core/api/data/datasources/connect_db_my_sql.dart';
+import 'package:technical_support_artphoto/core/api/data/repositories/technical_support_repo_impl.dart';
+import 'package:technical_support_artphoto/core/api/provider/providerModel.dart';
+import 'package:technical_support_artphoto/core/api/data/models/user.dart';
+import 'package:technical_support_artphoto/core/navigation/create_route.dart';
+import 'package:technical_support_artphoto/core/shared/logo_animate/draggable_logo.dart';
+import 'package:technical_support_artphoto/main.dart';
+import '../../../../core/utils/utils.dart';
+import '../../../../core/utils/utils.dart' as utils;
+
+class Authorization extends StatefulWidget {
+  const Authorization({super.key});
+
+  @override
+  State<Authorization> createState() => _AuthorizationState();
+}
+
+class _AuthorizationState extends State<Authorization> {
+  final passwordController = TextEditingController();
+  final numberFormatter = FilteringTextInputFormatter.allow(
+    RegExp(r'[0-9]'),
+  );
+  bool _hideText = true;
+
+  @override
+  void dispose() {
+    passwordController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final providerModel = Provider.of<ProviderModel>(context);
+    MyColor myColor = MyColor();
+    String? version = utils.packageInfo?.version;
+    double width = MediaQuery.sizeOf(context).width;
+    double height = MediaQuery.sizeOf(context).height;
+
+    return Container(
+      decoration: const BoxDecoration(
+          gradient: LinearGradient(
+              colors: [Colors.blue, Colors.purple], begin: Alignment.topRight, end: Alignment.bottomLeft)),
+      child: Scaffold(
+        backgroundColor: Colors.transparent,
+        appBar: AppBar(
+          flexibleSpace: myColor.appBar(),
+          title: Row(children: [
+            Text('v. $version', style: const TextStyle(fontSize: 12, color: Colors.white)),
+            const Expanded(
+                child: Center(child: Padding(padding: EdgeInsets.only(right: 45), child: Text('Авторизация')))),
+          ]),
+          centerTitle: true,
+        ),
+        body: SingleChildScrollView(
+          child: Column(
+            // alignment: AlignmentDirectional.center,
+            children: [
+              Padding(
+                padding: const EdgeInsets.only(top: 50, bottom: 30),
+                child: SizedBox(
+                    width: double.infinity,
+                    height: 200,
+                    child: DraggableLogo(
+                      child1: Image.asset('assets/logo/girl/girl2.png'),
+                      child2: Hero(
+                          tag: 'logo_hero',
+                          child: Image.asset(
+                            'assets/logo/logo.png',
+                            height: 200,
+                            width: 200,
+                          )),
+                      child3: Image.asset('assets/logo/girl/girl1.png'),
+                    )),
+              ),
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 40),
+                child: Container(
+                    padding: EdgeInsets.symmetric(horizontal: 20),
+                    decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(20),
+                        color: Colors.grey.shade50,
+                        boxShadow: const [BoxShadow(blurRadius: 10, color: Colors.black54, offset: Offset(0, 6))]),
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        Padding(
+                          padding: EdgeInsets.symmetric(horizontal: 10, vertical: 30),
+                          child: TextField(
+                            controller: passwordController,
+                            keyboardType: TextInputType.number,
+                            inputFormatters: [numberFormatter],
+                            obscureText: _hideText,
+                            obscuringCharacter: '*',
+                            autofocus: true,
+                            decoration: InputDecoration(
+                                border: OutlineInputBorder(
+                                  borderSide: BorderSide.none,
+                                  borderRadius: BorderRadius.circular(30.0),
+                                ),
+                                filled: true,
+                                fillColor: Colors.grey.shade200,
+                                prefixIcon: Icon(Icons.lock),
+                                suffixIcon: IconButton(
+                                    onPressed: () {
+                                      setState(() {
+                                        _hideText = !_hideText;
+                                      });
+                                    },
+                                    icon: Icon(Icons.remove_red_eye)),
+                                hintText: 'Пароль'),
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.all(28.0),
+                          child: SizedBox(
+                            width: MediaQuery.of(context).size.width,
+                            height: 50,
+                            child: ElevatedButton(
+                              onPressed: () {
+                                String password = passwordController.text;
+                                if (password.isNotEmpty) {
+                                  TechnicalSupportRepoImpl.downloadData.getUser(password).then((user){
+                                    if (user != null) {
+                                      providerModel.user[user.name] = user.access;
+                                      providerModel.user.remove('user');
+                                      Navigator.pushReplacement(
+                                          context, createRouteSlideTransition(const ArtphotoTech()));
+                                    } else {
+                                      passwordController.clear();
+                                      _showDialogNotValidationUser();
+                                    }
+                                  });
+                                }
+                              },
+                              style: ElevatedButton.styleFrom(
+                                  shape: RoundedRectangleBorder(
+                                borderRadius: BorderRadius.circular(20.0),
+                              )),
+                              child: const Text(
+                                'Вход',
+                                style: TextStyle(color: Colors.white, fontSize: 22),
+                              ),
+                            ),
+                          ),
+                        ),
+                      ],
+                    )),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  void _showDialogNotValidationUser() async {
+    return showDialog<void>(
+      context: context,
+      // barrierDismissible: false, // user must tap button!
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text('Пользователь не найден'),
+          content: SingleChildScrollView(
+            child: ListBody(
+              children: const <Widget>[
+                Text(
+                  'Введите пароль еще раз',
+                  textAlign: TextAlign.center,
+                ),
+              ],
+            ),
+          ),
+          shape: RoundedRectangleBorder(borderRadius: BorderRadius.all(Radius.circular(30))),
+          actions: <Widget>[
+            TextButton(
+              child: const Text('Закрыть'),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
