+++ lib/features/technics/TechnicViewAndChange.dart
