+++ lib/features/technics/data/models/technic_model.dart
@@ -1,0 +1,71 @@
+class TechnicModel implements Comparable {
+  int? id;
+  int? internalID;
+  String name = "";
+  String category = "";
+  int? cost;
+  String dateBuyTechnic = "";
+  String status = "";
+  String dislocation = "";
+  String dateChangeStatus = '';
+  String comment = "";
+  String testDriveDislocation = "";
+  String dateStartTestDrive = '';
+  String dateFinishTestDrive = '';
+  String resultTestDrive = '';
+  bool checkboxTestDrive = false;
+  String userTestDrive = '';
+  int? idTestDrive;
+
+  TechnicModel(
+      this.id,
+      this.internalID,
+      this.name,
+      this.category,
+      this.cost,
+      this.dateBuyTechnic,
+      this.status,
+      this.dislocation,
+      this.dateChangeStatus,
+      this.comment,
+      this.testDriveDislocation,
+      this.dateStartTestDrive,
+      this.dateFinishTestDrive,
+      this.resultTestDrive,
+      this.checkboxTestDrive);
+
+  TechnicModel.withoutTestDrive(
+      this.id,
+      this.internalID,
+      this.name,
+      this.category,
+      this.cost,
+      this.dateBuyTechnic,
+      this.status,
+      this.dislocation,
+      this.dateChangeStatus,
+      this.comment);
+
+  TechnicModel.testDrive(
+      this.internalID,
+      this.category,
+      this.testDriveDislocation,
+      this.dateStartTestDrive,
+      this.dateFinishTestDrive,
+      this.resultTestDrive,
+      this.checkboxTestDrive,
+      this.userTestDrive);
+
+  @override
+  String toString() {
+    return "{ id=$id, internalID=$internalID, name=$name, category=$category, coast=$cost, dateCoast=$dateBuyTechnic, status=$status, dislocation=$dislocation, comment=$comment }";
+  }
+
+  @override
+  int compareTo(other) {
+    int dislocationDiference = dislocation.compareTo(other.dislocation);
+    return dislocationDiference != 0
+        ? dislocationDiference
+        : category.compareTo(other.category);
+  }
+}
